split media-layout-context tests into a separate file and extract helper functions into test-util.js
