+++ packages/wonder-blocks-layout/components/media-layout.test.js
@@ -5,63 +5,7 @@ import {View} from "@khanacademy/wonder-blocks-core";
 
 import {mount, unmountAll} from "../../../utils/testing/mount.js";
 import MediaLayout from "./media-layout.js";
-import MediaLayoutContext from "./media-layout-context.js";
-import {
-    MEDIA_DEFAULT_SPEC,
-    MEDIA_INTERNAL_SPEC,
-    MEDIA_MODAL_SPEC,
-} from "../util/specs.js";
-import type {MediaSize} from "../util/types.js";
-
-const resizeWindow = (size: MediaSize) => {
-    if (size === "large") {
-        window.innerWidth = 1200;
-    } else if (size === "medium") {
-        window.innerWidth = 800;
-    } else if (size === "small") {
-        window.innerWidth = 640;
-    }
-
-    window.dispatchEvent(new Event("resize"));
-};
-
-const queries = {
-    small: "(max-width: 767px)",
-    medium: "(min-width: 768px) and (max-width: 1023px)",
-    large: "(min-width: 1024px)",
-};
-
-const checkQuery = (query, width) => {
-    return (
-        (width < 768 && query === queries.small) ||
-        (width < 1024 && query === queries.medium) ||
-        query === queries.large
-    );
-};
-
-const matchMedia = (query) => {
-    const listeners = new Set();
-
-    window.addEventListener("resize", () => {
-        for (const listener of listeners) {
-            listener({
-                matches: checkQuery(query, window.innerWidth),
-            });
-        }
-    });
-
-    return {
-        get matches() {
-            return checkQuery(query, window.innerWidth);
-        },
-        addListener: (listener) => {
-            listeners.add(listener);
-        },
-        removeListener: (listener) => {
-            listeners.delete(listener);
-        },
-    };
-};
+import {resizeWindow, matchMedia} from "../util/test-util.js";
 
 const getArgsToChildren = async (styleSheets) =>
     new Promise((resolve, reject) => {
@@ -94,149 +38,6 @@ describe("MediaLayout", () => {
     beforeEach(() => {
         unmountAll();
         window.matchMedia = matchMedia;
-    });
-
-    describe("overrideSize", () => {
-        it("should override the currentSize", async () => {
-            // Arrange
-            resizeWindow("large");
-            const promise = new Promise((resolve, reject) => {
-                mount(
-                    <MediaLayoutContext.Provider
-                        value={{
-                            overrideSize: "small",
-                            ssrSize: "large",
-                            mediaSpec: MEDIA_DEFAULT_SPEC,
-                        }}
-                    >
-                        <MediaLayout styleSheets={{}}>
-                            {({mediaSize, mediaSpec, styles}) => {
-                                resolve({mediaSize, mediaSpec, styles});
-                                return <View>Hello, world!</View>;
-                            }}
-                        </MediaLayout>
-                    </MediaLayoutContext.Provider>,
-                );
-            });
-
-            // Act
-            const args = await promise;
-
-            // Assert
-            expect(args.mediaSize).toEqual("small");
-        });
-    });
-
-    describe("ssrSize", () => {
-        it("should use the default ssrSize on the server", async () => {
-            // Arrange
-            delete window.matchMedia;
-            const promise = new Promise((resolve, reject) => {
-                mount(
-                    <MediaLayout styleSheets={{}}>
-                        {({mediaSize, mediaSpec, styles}) => {
-                            resolve({mediaSize, mediaSpec, styles});
-                            return <View>Hello, world!</View>;
-                        }}
-                    </MediaLayout>,
-                );
-            });
-
-            // Act
-            const args = await promise;
-
-            // Assert
-            expect(args.mediaSize).toEqual("large");
-        });
-
-        it("should use the provided ssrSize on the server", async () => {
-            // Arrange
-            delete window.matchMedia;
-            const promise = new Promise((resolve, reject) => {
-                mount(
-                    <MediaLayoutContext.Provider
-                        value={{
-                            overrideSize: undefined,
-                            ssrSize: "small",
-                            mediaSpec: MEDIA_DEFAULT_SPEC,
-                        }}
-                    >
-                        <MediaLayout styleSheets={{}}>
-                            {({mediaSize, mediaSpec, styles}) => {
-                                resolve({mediaSize, mediaSpec, styles});
-                                return <View>Hello, world!</View>;
-                            }}
-                        </MediaLayout>
-                    </MediaLayoutContext.Provider>,
-                );
-            });
-
-            // Act
-            const args = await promise;
-
-            // Assert
-            expect(args.mediaSize).toEqual("small");
-        });
-    });
-
-    describe("mediaSpec", () => {
-        it("MEDIA_INTERNAL_SPEC is always large", async () => {
-            // Arrange
-            resizeWindow("small");
-            const promise = new Promise((resolve, reject) => {
-                mount(
-                    <MediaLayoutContext.Provider
-                        value={{
-                            overrideSize: undefined,
-                            ssrSize: "small",
-                            mediaSpec: MEDIA_INTERNAL_SPEC,
-                        }}
-                    >
-                        <MediaLayout styleSheets={{}}>
-                            {({mediaSize, mediaSpec, styles}) => {
-                                resolve({mediaSize, mediaSpec, styles});
-                                return <View>Hello, world!</View>;
-                            }}
-                        </MediaLayout>
-                    </MediaLayoutContext.Provider>,
-                );
-            });
-
-            // Act
-            const args = await promise;
-
-            // Assert
-            expect(args.mediaSize).toEqual("large");
-        });
-
-        it("MEDIA_MODAL_SPEC is not medium", async () => {
-            // Arrange
-            resizeWindow("medium");
-            const promise = new Promise((resolve, reject) => {
-                mount(
-                    <MediaLayoutContext.Provider
-                        value={{
-                            overrideSize: undefined,
-                            ssrSize: "small",
-                            mediaSpec: MEDIA_MODAL_SPEC,
-                        }}
-                    >
-                        <MediaLayout styleSheets={{}}>
-                            {({mediaSize, mediaSpec, styles}) => {
-                                resolve({mediaSize, mediaSpec, styles});
-                                return <View>Hello, world!</View>;
-                            }}
-                        </MediaLayout>
-                    </MediaLayoutContext.Provider>,
-                );
-            });
-
-            // Act
-            const args = await promise;
-
-            // Assert
-            expect(args.mediaSize).toEqual("large");
-        });
     });
 
     describe("mediaSize", () => {
